verification complete
verified with previous edition of Selection.h before the scan over the
pt,eta,E ranges were added in that the plots agreed. All event/track
selections were turned off for the verification. The difference in the
nTrkEvent distributions between the old result from 20180126 and
20180228 were a result of the fact that before tracks were selected
based on eta and pt criteria. In the check since we are not doing that
we allow in more events while also getting more higher multiplicity
events. This accounts for the difference in the plots. We will now go
ahead with the scan over the phase spaces with the event/track
selections turned on. We expect significant change. Additionally the
events are now split <100 GeV for the Zpole and >100GeV for the excess
from LEP2.